+++ editor/CHANGES.md
@@ -1,4 +1,5 @@
 #### Version: 0.3.4-SNAPSHOT
+- **Improved**: Double clicking entity in scene outline will center camera on it
 
 #### Version: 0.3.3
 - **Improved**: Improved performance when performing select all on more than 2000 entities